RF: Move base functionality to a base class and subclass it for each device
